+++ tracker/npm_package/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 - Add `logging` option
+- Improve `callback` option
 
 ## [0.2.2] - 2025-06-16